Write second draft of MCTS
This version uses an Int32Array heap and a custom game analyzer.
I'm hoping that allocating a vast majority of the memory (required for
the tree) ahead of time will improve the performance of each rollout
cycle.
Unfortunately, the heavy use of bitwise operators comes at the expense
of readability.