Rename getAnalyzer to getStateAnalyzer
This clarifies the distinction between the "dumb" state analyzer, which
doesn't provide any information on what the optimal action is, and the
MCTS analyzer, which does.

It's probably still confusing why there is a state analyzer and an
MCTS analyzer, so maybe I should explicitly rename the state analyzer
to "objective analyzer" or "non-speculative analyzer".